pyproject.toml
[project]
name = "personal-brain"
version = "0.1.0"
description = "A second brain assistant powered by Obsidian and Claude"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "you@example.com"}
]

dependencies = [
    # Core
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",

    # Data Layer
    "chromadb>=0.5",
    "watchfiles>=0.21",
    "mistune>=3.0",
    "python-frontmatter>=1.0",

    # Intelligence Layer
    "anthropic>=0.40",
    "openai>=1.0",
    "rank-bm25>=0.2",

    # API Layer
    "fastapi>=0.110",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",

    # CLI Layer
    "typer>=0.12",
    "rich>=13.0",

    # Scheduling
    "apscheduler>=3.10",

    # Web UI
    "jinja2>=3.1",

    # Notifications
    "plyer>=2.1",

    # Utilities
    "structlog>=24.0",
    "tenacity>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
brain = "brain.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/brain"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
